refactor(react-tinacms-editor): moves imageProp logic to Wysiwyg
