+++ packages/react-tinacms-editor/src/tinacms-inline/inline-wysiwyg.tsx
@@ -21,7 +21,7 @@ import { useCMS } from 'tinacms'
 import { InlineField, FocusRing } from 'react-tinacms-inline'
 import { FocusRingOptions } from 'react-tinacms-inline/src/styles'
 import { Wysiwyg } from '../components/Wysiwyg'
-import { EditorProps, ImageProps } from '../types'
+import { EditorProps } from '../types'
 
 export interface InlineWysiwygFieldProps extends Omit<EditorProps, 'input'> {
   name: string
@@ -33,41 +33,10 @@ export function InlineWysiwyg({
   name,
   children,
   focusRing = true,
-  imageProps: passedInImageProps,
+  imageProps,
   ...wysiwygProps
 }: InlineWysiwygFieldProps) {
   const cms = useCMS()
-
-  const imageProps: ImageProps | undefined = React.useMemo(() => {
-    if (!passedInImageProps) return
-    return {
-      async upload(files: File[]) {
-        const filesToUpload = files.map(file => ({
-          directory: passedInImageProps?.directory || '',
-          file,
-        }))
-
-        const allMedia = await cms.media.persist(filesToUpload)
-
-        return allMedia.map(media => {
-          if (passedInImageProps.parse) {
-            return passedInImageProps.parse(media)
-          } else {
-            return media.filename
-          }
-        })
-      },
-      previewSrc(src) {
-        return cms.media.previewSrc(src)
-      },
-      ...passedInImageProps,
-    }
-  }, [
-    cms.media.store,
-    passedInImageProps?.directory,
-    passedInImageProps?.previewSrc,
-    passedInImageProps?.upload,
-  ])
 
   if (cms.disabled) {
     return children
